scripts/requirements.txt
requests
beautifulsoup4
dnspython
httpx
tldextract
ollama
